Refactor: 프론트 오류 해결
member json구조로 변경 및 CORS OPTIONS 추가
