+++ src/main/java/gift/DTO/SuccessMessageDTO.java
@@ -1,0 +1,12 @@
+package gift.DTO;
+
+public class SuccessMessageDTO {
+   private String message;
+   public SuccessMessageDTO(String message){
+       this.message = message;
+   }
+
+    public String getMessage() {
+        return message;
+    }
+}
