+++ src/main/java/gift/Service/WebMvcConfig.java
@@ -1,6 +1,5 @@
-package gift.Config;
+package gift.Service;
 
-import gift.Service.BearerAuthInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
